separate about navigation classes
