+++ model/src/main/java/com/freshdigitable/udonroad2/model/app/navigation/NavigationDispatcher.kt
@@ -1,18 +1,12 @@
 package com.freshdigitable.udonroad2.model.app.navigation
 
-import androidx.appcompat.app.AppCompatActivity
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleObserver
 import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.OnLifecycleEvent
-import androidx.lifecycle.observe
+import androidx.lifecycle.distinctUntilChanged
 import androidx.lifecycle.toLiveData
 import com.freshdigitable.udonroad2.model.app.di.ActivityScope
 import io.reactivex.BackpressureStrategy
 import io.reactivex.Flowable
 import io.reactivex.Observable
-import io.reactivex.disposables.Disposable
 import io.reactivex.subjects.BehaviorSubject
 import io.reactivex.subjects.PublishSubject
 import timber.log.Timber
@@ -31,8 +25,6 @@ class NavigationDispatcher @Inject constructor() {
 typealias AppAction<T> = Observable<T>
 typealias AppViewState<T> = LiveData<T>
 
-data class StateHolder<T>(val value: T?)
-
 inline fun <T> NavigationDispatcher.toAction(
     block: PublishSubject<NavigationEvent>.() -> AppAction<T>
 ): AppAction<T> {
@@ -49,6 +41,7 @@ inline fun <T, reified E> AppAction<T>.toViewState(
     }
         .toFlowable(BackpressureStrategy.BUFFER)
         .toLiveData()
+        .distinctUntilChanged()
 }
 
 inline fun <reified T> Observable<NavigationEvent>.filterByType(): Observable<T> {
@@ -57,50 +50,4 @@ inline fun <reified T> Observable<NavigationEvent>.filterByType(): Observable<T>
 
 inline fun <reified T> Flowable<NavigationEvent>.filterByType(): Flowable<T> {
     return this.filter { it is T }.cast(T::class.java)
-}
-
-interface NavigationEvent
-
-interface FragmentContainerState
-
-interface ViewState {
-    val containerState: FragmentContainerState
-}
-
-sealed class CommonEvent : NavigationEvent {
-    data class Back(
-        val currentState: ViewState?,
-        val prevContainerState: FragmentContainerState? = null
-    ) : CommonEvent()
-}
-
-abstract class Navigation<T : FragmentContainerState>(
-    val navigator: NavigationDispatcher,
-    protected val activity: AppCompatActivity
-) : LifecycleObserver {
-
-    private val state: MutableLiveData<T?> = MutableLiveData()
-    protected val currentState: T?
-        get() = state.value
-
-    private val eventDisposable: Disposable
-
-    init {
-        activity.lifecycle.addObserver(this)
-        eventDisposable = navigator.emitter
-            .subscribe { event ->
-                val s = onEvent(event) ?: return@subscribe
-                state.postValue(s)
-            }
-
-        state.observe(activity) { navigate(it) }
-    }
-
-    abstract fun onEvent(event: NavigationEvent): T?
-    abstract fun navigate(s: T?)
-
-    @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
-    fun onDestroy() {
-        eventDisposable.dispose()
-    }
 }
