+++ model/src/main/java/com/freshdigitable/udonroad2/model/app/navigation/NavigationEvents.kt
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2020. Matsuda, Akihit (akihito104)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.freshdigitable.udonroad2.model.app.navigation
+
+interface NavigationEvent
+interface FragmentContainerState
+interface ViewState {
+    val containerState: FragmentContainerState
+}
+
+sealed class CommonEvent : NavigationEvent {
+    data class Back(
+        val currentState: ViewState?,
+        val prevContainerState: FragmentContainerState? = null
+    ) : CommonEvent()
+}
+
+data class StateHolder<T>(val value: T?)
